+++ lib/l10n/app_localizations.dart
@@ -97,7 +97,7 @@ abstract class AppLocalizations {
   /// The title of the application
   ///
   /// In en, this message translates to:
-  /// **'Companion GPT'**
+  /// **'NafsAI'**
   String get appTitle;
 
   /// Greeting for male users